refactor: 사용하지 않는 파일 정리, 테이블 column명 snake case로 변경
