+++ README.md
@@ -7,21 +7,21 @@
 - [x] 테이블 설계
 
 ```sql
-CREATE TABLE cars
-(
-`name`       VARCHAR(20) NOT NULL,
-`position`     INT       NOT NULL,
-`gameId`       INT       NOT NULL,
-PRIMARY KEY (`name`),
-FOREIGN KEY ( gameId ) REFERENCES games ( gameId )
-);
 CREATE TABLE games
 (
-`gameId`        INT          NOT NULL AUTO_INCREMENT,
-`count        ` INT          NOT NULL,
-`winner`        VARCHAR(100) NOT NULL,
-`dateTime`      DATETIME     NOT NULL,
-PRIMARY KEY (`gameId`)
+  `game_id`      INT           NOT NULL AUTO_INCREMENT,
+  `count`        INT           NOT NULL,
+  `winner`       VARCHAR(100)  NOT NULL,
+  `timeStamp`    TIMESTAMP     NOT NULL,
+  PRIMARY KEY (`game_id`)
+);
+CREATE TABLE cars
+(
+  `name`       VARCHAR(20) NOT NULL,
+  `position`     INT       NOT NULL,
+  `game_id`       INT       NOT NULL,
+  PRIMARY KEY (`name`, `game_id`),
+  FOREIGN KEY ( `game_id` ) REFERENCES `games` ( `game_id` )
 );
 
 ```
