--- src/main/java/racingcar/RacingCarApplication.java
@@ -1,13 +1,0 @@
-package racingcar;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RacingCarApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RacingCarApplication.class, args);
-	}
-
-}
