+++ src/main/resources/data.sql
@@ -1,16 +1,16 @@
 CREATE TABLE games
 (
-`gameId`        INT          NOT NULL AUTO_INCREMENT,
-`count`         INT          NOT NULL,
+`game_id`      INT           NOT NULL AUTO_INCREMENT,
+`count`        INT           NOT NULL,
 `winner`       VARCHAR(100)  NOT NULL,
 `timeStamp`    TIMESTAMP     NOT NULL,
-PRIMARY KEY (`gameId`)
+PRIMARY KEY (`game_id`)
 );
 CREATE TABLE cars
 (
 `name`       VARCHAR(20) NOT NULL,
 `position`     INT       NOT NULL,
-`gameId`       INT       NOT NULL,
-PRIMARY KEY (`name`, `gameId`),
-FOREIGN KEY ( `gameId` ) REFERENCES `games` ( `gameId` )
+`game_id`       INT       NOT NULL,
+PRIMARY KEY (`name`, `game_id`),
+FOREIGN KEY ( `game_id` ) REFERENCES `games` ( `game_id` )
 );
